@@ -8,13 +8,13 @@ import 'package:uuid/uuid.dart';
 
 part 'message.g.dart';
 
-class _PinExpires {
-  const _PinExpires();
+class _NullConst {
+  const _NullConst();
 }
 
-const _pinExpires = _PinExpires();
+const _nullConst = _NullConst();
 
-/// Enum defining the status of a sending message
+/// Enum defining the status of a sending message.
 enum MessageSendingStatus {
   /// Message is being sent
   sending,
@@ -40,10 +40,10 @@ enum MessageSendingStatus {
   sent,
 }
 
-/// The class that contains the information about a message
+/// The class that contains the information about a message.
 @JsonSerializable()
 class Message extends Equatable {
-  /// Constructor used for json serialization
+  /// Constructor used for json serialization.
   Message({
     String? id,
     this.text,
@@ -58,44 +58,47 @@ class Message extends Equatable {
     this.ownReactions,
     this.parentId,
     this.quotedMessage,
-    this.quotedMessageId,
+    String? quotedMessageId,
     this.replyCount = 0,
     this.threadParticipants,
     this.showInChannel,
     this.command,
     DateTime? createdAt,
     DateTime? updatedAt,
+    this.deletedAt,
     this.user,
     this.pinned = false,
     this.pinnedAt,
     DateTime? pinExpires,
     this.pinnedBy,
     this.extraData = const {},
-    this.deletedAt,
-    this.status = MessageSendingStatus.sent,
+    this.status = MessageSendingStatus.sending,
     this.i18n,
   })  : id = id ?? const Uuid().v4(),
         pinExpires = pinExpires?.toUtc(),
-        createdAt = createdAt ?? DateTime.now(),
-        updatedAt = updatedAt ?? DateTime.now();
+        _createdAt = createdAt,
+        _updatedAt = updatedAt,
+        _quotedMessageId = quotedMessageId;
 
-  /// Create a new instance from a json
+  /// Create a new instance from JSON.
   factory Message.fromJson(Map<String, dynamic> json) => _$MessageFromJson(
         Serializer.moveToExtraDataFromRoot(json, topLevelFields),
+      ).copyWith(
+        status: MessageSendingStatus.sent,
       );
 
   /// The message ID. This is either created by Stream or set client side when
   /// the message is added.
   final String id;
 
-  /// The text of this message
+  /// The text of this message.
   final String? text;
 
-  /// The status of a sending message
+  /// The status of a sending message.
   @JsonKey(ignore: true)
   final MessageSendingStatus status;
 
-  /// The message type
+  /// The message type.
   @JsonKey(
     includeIfNull: false,
     toJson: Serializer.readOnly,
@@ -107,15 +110,15 @@ class Message extends Equatable {
   @JsonKey(includeIfNull: false)
   final List<Attachment> attachments;
 
-  /// The list of user mentioned in the message
+  /// The list of user mentioned in the message.
   @JsonKey(toJson: User.toIds)
   final List<User> mentionedUsers;
 
-  /// A map describing the count of number of every reaction
+  /// A map describing the count of number of every reaction.
   @JsonKey(includeIfNull: false, toJson: Serializer.readOnly)
   final Map<String, int>? reactionCounts;
 
-  /// A map describing the count of score of every reaction
+  /// A map describing the count of score of every reaction.
   @JsonKey(includeIfNull: false, toJson: Serializer.readOnly)
   final Map<String, int>? reactionScores;
 
@@ -130,12 +133,14 @@ class Message extends Equatable {
   /// The ID of the parent message, if the message is a thread reply.
   final String? parentId;
 
-  /// A quoted reply message
+  /// A quoted reply message.
   @JsonKey(toJson: Serializer.readOnly)
   final Message? quotedMessage;
 
+  final String? _quotedMessageId;
+
   /// The ID of the quoted message, if the message is a quoted reply.
-  final String? quotedMessageId;
+  String? get quotedMessageId => _quotedMessageId ?? quotedMessage?.id;
 
   /// Reserved field indicating the number of replies for this message.
   @JsonKey(includeIfNull: false, toJson: Serializer.readOnly)
@@ -148,10 +153,10 @@ class Message extends Equatable {
   /// Check if this message needs to show in the channel.
   final bool? showInChannel;
 
-  /// If true the message is silent
+  /// If true the message is silent.
   final bool silent;
 
-  /// If true the message is shadowed
+  /// If true the message is shadowed.
   @JsonKey(
     includeIfNull: false,
     toJson: Serializer.readOnly,
@@ -162,56 +167,61 @@ class Message extends Equatable {
   @JsonKey(includeIfNull: false, toJson: Serializer.readOnly)
   final String? command;
 
-  /// Reserved field indicating when the message was created.
-  @JsonKey(includeIfNull: false, toJson: Serializer.readOnly)
-  final DateTime createdAt;
-
-  /// Reserved field indicating when the message was updated last time.
-  @JsonKey(includeIfNull: false, toJson: Serializer.readOnly)
-  final DateTime updatedAt;
-
-  /// User who sent the message
-  @JsonKey(includeIfNull: false, toJson: Serializer.readOnly)
-  final User? user;
-
-  /// If true the message is pinned
-  final bool pinned;
-
-  /// Reserved field indicating when the message was pinned
-  @JsonKey(toJson: Serializer.readOnly)
-  final DateTime? pinnedAt;
-
-  /// Reserved field indicating when the message will expire
-  ///
-  /// if `null` message has no expiry
-  final DateTime? pinExpires;
-
-  /// Reserved field indicating who pinned the message
-  @JsonKey(toJson: Serializer.readOnly)
-  final User? pinnedBy;
-
-  /// Message custom extraData
-  @JsonKey(includeIfNull: false)
-  final Map<String, Object?> extraData;
-
-  /// True if the message is a system info
-  bool get isSystem => type == 'system';
-
-  /// True if the message has been deleted
-  bool get isDeleted => type == 'deleted';
-
-  /// True if the message is ephemeral
-  bool get isEphemeral => type == 'ephemeral';
+  final DateTime? _createdAt;
 
   /// Reserved field indicating when the message was deleted.
   @JsonKey(includeIfNull: false, toJson: Serializer.readOnly)
   final DateTime? deletedAt;
+
+  /// Reserved field indicating when the message was created.
+  @JsonKey(includeIfNull: false, toJson: Serializer.readOnly)
+  DateTime get createdAt => _createdAt ?? DateTime.now();
+
+  final DateTime? _updatedAt;
+
+  /// Reserved field indicating when the message was updated last time.
+  @JsonKey(includeIfNull: false, toJson: Serializer.readOnly)
+  DateTime get updatedAt => _updatedAt ?? DateTime.now();
+
+  /// User who sent the message.
+  @JsonKey(includeIfNull: false, toJson: Serializer.readOnly)
+  final User? user;
+
+  /// If true the message is pinned.
+  final bool pinned;
+
+  /// Reserved field indicating when the message was pinned.
+  @JsonKey(toJson: Serializer.readOnly)
+  final DateTime? pinnedAt;
+
+  /// Reserved field indicating when the message will expire.
+  ///
+  /// If `null` message has no expiry.
+  final DateTime? pinExpires;
+
+  /// Reserved field indicating who pinned the message.
+  @JsonKey(toJson: Serializer.readOnly)
+  final User? pinnedBy;
+
+  /// Message custom extraData.
+  @JsonKey(includeIfNull: false)
+  final Map<String, Object?> extraData;
+
+  /// True if the message is a system info.
+  bool get isSystem => type == 'system';
+
+  /// True if the message has been deleted.
+  bool get isDeleted => type == 'deleted';
+
+  /// True if the message is ephemeral.
+  bool get isEphemeral => type == 'ephemeral';
 
   /// A Map of translations.
   @JsonKey(includeIfNull: false)
   final Map<String, String>? i18n;
 
   /// Known top level fields.
+  ///
   /// Useful for [Serializer] methods.
   static const topLevelFields = [
     'id',
@@ -244,7 +254,7 @@ class Message extends Equatable {
     'i18n',
   ];
 
-  /// Serialize to json
+  /// Serialize to json.
   Map<String, dynamic> toJson() => Serializer.moveFromExtraDataToRoot(
         _$MessageToJson(this),
       );
@@ -256,18 +266,18 @@ class Message extends Equatable {
     String? type,
     List<Attachment>? attachments,
     List<User>? mentionedUsers,
+    bool? silent,
+    bool? shadowed,
     Map<String, int>? reactionCounts,
     Map<String, int>? reactionScores,
     List<Reaction>? latestReactions,
     List<Reaction>? ownReactions,
     String? parentId,
-    Message? quotedMessage,
-    String? quotedMessageId,
+    Object? quotedMessage = _nullConst,
+    Object? quotedMessageId = _nullConst,
     int? replyCount,
     List<User>? threadParticipants,
     bool? showInChannel,
-    bool? shadowed,
-    bool? silent,
     String? command,
     DateTime? createdAt,
     DateTime? updatedAt,
@@ -275,7 +285,7 @@ class Message extends Equatable {
     User? user,
     bool? pinned,
     DateTime? pinnedAt,
-    Object? pinExpires = _pinExpires,
+    Object? pinExpires = _nullConst,
     User? pinnedBy,
     Map<String, Object?>? extraData,
     MessageSendingStatus? status,
@@ -284,41 +294,68 @@ class Message extends Equatable {
     assert(() {
       if (pinExpires is! DateTime &&
           pinExpires != null &&
-          pinExpires is! _PinExpires) {
+          pinExpires is! _NullConst) {
         throw ArgumentError('`pinExpires` can only be set as DateTime or null');
       }
       return true;
     }(), 'Validate type for pinExpires');
+
+    assert(() {
+      if (quotedMessage is! Message &&
+          quotedMessage != null &&
+          quotedMessage is! _NullConst) {
+        throw ArgumentError(
+          '`quotedMessage` can only be set as Message or null',
+        );
+      }
+      return true;
+    }(), 'Validate type for quotedMessage');
+
+    assert(() {
+      if (quotedMessageId is! String &&
+          quotedMessageId != null &&
+          quotedMessageId is! _NullConst) {
+        throw ArgumentError(
+          '`quotedMessage` can only be set as String or null',
+        );
+      }
+      return true;
+    }(), 'Validate type for quotedMessage');
+
     return Message(
       id: id ?? this.id,
       text: text ?? this.text,
       type: type ?? this.type,
       attachments: attachments ?? this.attachments,
       mentionedUsers: mentionedUsers ?? this.mentionedUsers,
+      silent: silent ?? this.silent,
+      shadowed: shadowed ?? this.shadowed,
       reactionCounts: reactionCounts ?? this.reactionCounts,
       reactionScores: reactionScores ?? this.reactionScores,
       latestReactions: latestReactions ?? this.latestReactions,
       ownReactions: ownReactions ?? this.ownReactions,
       parentId: parentId ?? this.parentId,
-      quotedMessage: quotedMessage ?? this.quotedMessage,
-      quotedMessageId: quotedMessageId ?? this.quotedMessageId,
+      quotedMessage: quotedMessage == _nullConst
+          ? this.quotedMessage
+          : quotedMessage as Message?,
+      quotedMessageId: quotedMessageId == _nullConst
+          ? _quotedMessageId
+          : quotedMessageId as String?,
       replyCount: replyCount ?? this.replyCount,
       threadParticipants: threadParticipants ?? this.threadParticipants,
       showInChannel: showInChannel ?? this.showInChannel,
       command: command ?? this.command,
-      createdAt: createdAt ?? this.createdAt,
-      silent: silent ?? this.silent,
-      extraData: extraData ?? this.extraData,
-      user: user ?? this.user,
-      shadowed: shadowed ?? this.shadowed,
-      updatedAt: updatedAt ?? this.updatedAt,
+      createdAt: createdAt ?? _createdAt,
+      updatedAt: updatedAt ?? _updatedAt,
       deletedAt: deletedAt ?? this.deletedAt,
-      status: status ?? this.status,
+      user: user ?? this.user,
       pinned: pinned ?? this.pinned,
       pinnedAt: pinnedAt ?? this.pinnedAt,
-      pinnedBy: pinnedBy ?? this.pinnedBy,
       pinExpires:
-          pinExpires == _pinExpires ? this.pinExpires : pinExpires as DateTime?,
+          pinExpires == _nullConst ? this.pinExpires : pinExpires as DateTime?,
+      pinnedBy: pinnedBy ?? this.pinnedBy,
+      extraData: extraData ?? this.extraData,
+      status: status ?? this.status,
       i18n: i18n ?? this.i18n,
     );
   }
@@ -331,6 +368,8 @@ class Message extends Equatable {
         type: other.type,
         attachments: other.attachments,
         mentionedUsers: other.mentionedUsers,
+        silent: other.silent,
+        shadowed: other.shadowed,
         reactionCounts: other.reactionCounts,
         reactionScores: other.reactionScores,
         latestReactions: other.latestReactions,
@@ -343,17 +382,15 @@ class Message extends Equatable {
         showInChannel: other.showInChannel,
         command: other.command,
         createdAt: other.createdAt,
-        silent: other.silent,
-        extraData: other.extraData,
-        user: other.user,
-        shadowed: other.shadowed,
         updatedAt: other.updatedAt,
         deletedAt: other.deletedAt,
-        status: other.status,
+        user: other.user,
         pinned: other.pinned,
         pinnedAt: other.pinnedAt,
         pinExpires: other.pinExpires,
         pinnedBy: other.pinnedBy,
+        extraData: other.extraData,
+        status: other.status,
         i18n: other.i18n,
       );
 
@@ -377,8 +414,8 @@ class Message extends Equatable {
         shadowed,
         silent,
         command,
-        createdAt,
-        updatedAt,
+        _createdAt,
+        _updatedAt,
         deletedAt,
         user,
         pinned,
