@@ -634,7 +634,9 @@ class _MessageActionsModalState extends State<MessageActionsModal> {
                 widget.editMessageInputBuilder!(context, widget.message)
               else
                 MessageInput(
-                  editMessage: widget.message,
+                  messageInputController: MessageInputController(
+                    message: widget.message,
+                  ),
                   preMessageSending: (m) {
                     FocusScope.of(context).unfocus();
                     Navigator.pop(context);
