@@ -23,7 +23,11 @@ export 'src/localization/stream_chat_localizations.dart';
 export 'src/localization/translations.dart' show DefaultTranslations;
 export 'src/mention_tile.dart';
 export 'src/message_action.dart';
-export 'src/message_input.dart';
+export 'src/message_input/countdown_button.dart';
+export 'src/message_input/message_input.dart';
+export 'src/message_input/stream_attachment_picker.dart';
+export 'src/message_input/stream_message_send_button.dart';
+export 'src/message_input/stream_message_text_field.dart';
 export 'src/message_list_view.dart';
 export 'src/message_search_item.dart';
 export 'src/message_search_list_view.dart';
